feature: move register end points to IdentityController and remove unnecessary classes.
