--- identity/src/main/kotlin/net/thechance/identity/exception/RegisterException.kt
@@ -1,5 +1,0 @@
-package net.thechance.identity.exception
-
-abstract class RegisterException(message: String): Exception(message)
-
-class UserAlreadyExistsException(message: String = "User already exists"): RegisterException(message)
